+++ test/decorators-unsure-fixtures/test.mjs
@@ -1,17 +1,38 @@
-import { test, getTestName, execPlaywrightTestWithError, DEFAULT_CMD } from '../helpers.mjs';
+import { test, TestDir, execPlaywrightTestWithError, DEFAULT_CMD } from '../helpers.mjs';
 
-test(
-  getTestName(import.meta),
-  (t) =>
-    execPlaywrightTestWithError(
-      t.name,
-      [
-        `Can't guess fixture for decorator step "BasePage: step" in file: features/guess.feature`,
-        `Please refactor your Page Object classes or set one of the following tags: @fixture:todoPage, @fixture:todoPage2`,
-      ],
-      `FEATURE=guess.feature ${DEFAULT_CMD}`,
-    ),
+const testDir = new TestDir(import.meta);
 
-  // todo:
-  // execPlaywrightTestWithError(t.name, [`sgsedg`], `FEATURE=scenario-tag.feature ${DEFAULT_CMD}`),
+test(testDir.name + ' (guess)', () =>
+  execPlaywrightTestWithError(
+    testDir.name,
+    [
+      `Can't guess fixture for decorator step "BasePage: step" in file: features/guess.feature`,
+      `Please refactor your Page Object classes or set one of the following tags: @fixture:todoPage, @fixture:todoPage2`,
+    ],
+    `FEATURE=guess.feature ${DEFAULT_CMD}`,
+  ),
+);
+
+test(testDir.name + ' (scenario-tag)', () =>
+  execPlaywrightTestWithError(
+    testDir.name,
+    [`contains 1 step(s) not compatible with required fixture "todoPage"`],
+    `FEATURE=scenario-tag.feature ${DEFAULT_CMD}`,
+  ),
+);
+
+test(testDir.name + ' (feature-tag)', () =>
+  execPlaywrightTestWithError(
+    testDir.name,
+    [`contains 1 step(s) not compatible with required fixture "todoPage"`],
+    `FEATURE=feature-tag.feature ${DEFAULT_CMD}`,
+  ),
+);
+
+test(testDir.name + ' (bg)', () =>
+  execPlaywrightTestWithError(
+    testDir.name,
+    [`Scenario "Background" contains 1 step(s) not compatible with required fixture "todoPage"`],
+    `FEATURE=bg.feature ${DEFAULT_CMD}`,
+  ),
 );